% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGR.R
\name{CAGR}
\alias{CAGR}
\title{# Compound Annual Growth Rate}
\usage{
CAGR(x, m)
}
\arguments{
\item{x}{xts object}

\item{m}{number to annualize (for monthly data, it will be 12, for dayli 256 ect.)}
}
\value{
Returns the Compound Annual Growth Rate
}
\description{
# Compound Annual Growth Rate
}
