% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrMov.R
\name{ErrMov}
\alias{ErrMov}
\title{Error adjusted moving average}
\usage{
ErrMov(Price_Daily, SMAperiod = 10, Lag = 3, CompareObj = NULL,
  period = c("days", "weeks", "months", "quarters", "years"))
}
\arguments{
\item{Price_Daily}{xts object in form of a dataframe with daily prices}

\item{SMAperiod}{Period in which one want to adjust prices. Hence this period is used when calculating forecast and thereby errors}

\item{Lag}{Number of periods used when calculating simple momentum of the roc/errors}

\item{CompareObj}{Object of which final object should be compare to. Ex. if period is monthly, then a frame with the wanted index should be provided}

\item{period}{Periods at which one is rebalancing. Can be "days", "weeks", "months", "quarters" or "years"}
}
\value{
Matrix of simple moving average o
}
\description{
Error adjusted moving average
}
