% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_RelROCMom.R
\name{fun_RelROCMom}
\alias{fun_RelROCMom}
\title{Relative ROC based Momentum}
\usage{
fun_RelROCMom(listData, period = c("days", "weeks", "months", "quarters",
  "years"), lookback = c(3, 6, 12), dateStart = NULL,
  weights = c(1/6, 2/3, 1/6), num.assets = 8, TC = NULL,
  savePlot = NULL, name = NULL, width = 8, height = 6)
}
\arguments{
\item{listData}{Data from Yahoo in form OHLCAV. The adjusted prices is used s these must be present}

\item{period}{Period at which one want to obtain data. Can be "days", "weeks", "months", "quarters" or "years"}

\item{lookback}{Lookback periods used to rank assets}

\item{dateStart}{Date at which the strategy shall start. The lookback period is looking before this date so it is the start date of the strategy. The form should be ex: "2018/", "2018-01/" or "2018-01-01/"}

\item{weights}{weights when using a weighted momentum. Vector which should have lenth three and sum to one}

\item{num.assets}{Number of assets in optimal portfolio}

\item{TC}{Tranding costs. If null, then vector without trading cost will be returned}

\item{savePlot}{Outdir if one want to save plot}

\item{name}{If savePlot !is.null then the name (OBS: also _MOM_period is appended)}

\item{width}{If savePlot !is.null then the width of plot}

\item{height}{If savePlot !is.null then the height of plot}
}
\value{
list
 \item{name plotCumReturn}{ggplot of cumreturns for different assets}
 \item{name StatTable}{Table of statistics}
 \item{name xtable_StatTable}{LateX code for above table of statistics}
}
\description{
Returns a plot and statistics for a relative momentum based on ROC ranking
}
