% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Shiny_PropAbs.R
\name{fun_Shiny_PropAbs}
\alias{fun_Shiny_PropAbs}
\title{fun_Shiny_PropAbs}
\usage{
fun_Shiny_PropAbs(Assets, ChoosenAssets = NULL, rf.asset = 1,
  k = 200, cf = 0.6, dateStart = NULL, dateEnd = NULL,
  TCrate = 0.001, TCmin = 29, size = 1, nd = 256, alpha = 0.05,
  UseTrans = TRUE)
}
\arguments{
\item{Assets}{List of assets}

\item{ChoosenAssets}{Assets one want to look at}

\item{rf.asset}{Risk free rate}

\item{k}{Number of lookback periods}

\item{cf}{confidence interval}

\item{dateStart}{Start date of strategy}

\item{dateEnd}{End date of strategy}

\item{TCrate}{Transaction cost given rate}

\item{TCmin}{Transactin cost minimum amount}

\item{size}{Size of each transaction}

\item{nd}{Number of days in a year}

\item{alpha}{alpha value used for calculating VaR and ES}

\item{UseTrans}{Boolean if one want to use tranaction costs}
}
\value{
list
 \item{name weights}{Table of position of weights}
 \item{name table}{Table of statistics}
 \item{name cumplot}{Plot of cumulated returns}
 \item{name weightplot}{Plot of position of weights}
}
\description{
Returns plot for cumulated returns and weights, table of statistics and table of weights.
Designed to be used fr IPA's Shiny App
}
