% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Shiny_RelErrMom.R
\name{fun_Shiny_RelErrMom}
\alias{fun_Shiny_RelErrMom}
\title{Relative Error Adjusted based Momentum}
\usage{
fun_Shiny_RelErrMom(listData, period = c("days", "weeks", "months",
  "quarters", "years"), lookback = c(3, 6, 12), dateStart = NULL,
  dateEnd = NULL, weights = c(1/6, 2/3, 1/6), NumAssets = 8,
  size = NULL, UseTC = TRUE, TCmin = 30, TCrate = 0.001,
  SMAperiod = 10, ChoosenAssets = NULL)
}
\arguments{
\item{listData}{Data from Yahoo in form OHLCAV. The adjusted prices is used s these must be present}

\item{period}{Period at which one want to obtain data. Can be "days", "weeks", "months", "quarters" or "years"}

\item{lookback}{Lookback periods used to rank assets}

\item{dateStart}{Date at which the strategy shall start. The lookback period is looking before this date so it is the start date of the strategy.}

\item{dateEnd}{Date at which the strategy shall end}

\item{weights}{weights when using a weighted momentum. Vector which should have lenth three and sum to one}

\item{NumAssets}{Number of assets in optimal portfolio}

\item{size}{Number of asset one buy (or sell) in each transaction. Default is one}

\item{UseTC}{Logical if one want to use transaction cast}

\item{TCmin}{Minimum transactionamount}

\item{TCrate}{Rate in which the transaction cost is calculated. This is only used if it is over the minimal amount}

\item{SMAperiod}{int Period one is using to calculate errors. These are calcualted on daily frequency and not the period specified}

\item{ChoosenAssets}{Assets one have choosen in strategy}
}
\value{
list
 \item{name plotCumReturn}{ggplot of cumreturns for different assets}
 \item{name StatTable}{Table of statistics}
 \item{name rank.first}{Table of rank for first lookback period}
 \item{name rank.second}{Table of rank for second lookback period}
 \item{name rank.third}{Table of rank for third lookback period}
 \item{name rank.ave}{Table of rank for average lookback period}
 \item{name rank.weight.ave}{Table of rank for weighted average lookback period}
}
\description{
Returns a plot and statistics for a relative momentum based on error adjusted ranking
OBS! Objects with only NA's values will be removed under calculations.
This function is customized to work with IPA's Shiny App
}
