% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_Ad.R
\name{period_Ad}
\alias{period_Ad}
\title{Period Adjusted prices}
\usage{
period_Ad(x, period = c("days", "weeks", "months", "quarters", "years"))
}
\arguments{
\item{x}{xts object in the form OHLCAV}

\item{period}{Period at which one want to obtain data. Can be "days", "weeks", "months", "quarters" or "years"}
}
\value{
Adjusted prices
}
\description{
Period Adjusted prices
}
