% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankRB.R
\name{RankRB}
\alias{RankRB}
\title{Rank returns}
\usage{
RankRB(x)
}
\arguments{
\item{x}{Returns in a dataframe or matrix.}
}
\value{
Object with input frame ranked
}
\description{
Rank assets given thier return, ROC or other format the input is given in
}
