% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeMomentum.R
\name{RelativeMomentum}
\alias{RelativeMomentum}
\title{Relative Momentum Strategy}
\usage{
RelativeMomentum(xts.ret, xts.rank, n = 1, ret.fill.na = 3, TC = 30)
}
\arguments{
\item{xts.ret}{Object of returns}

\item{xts.rank}{Object of ranked returns}

\item{n}{Number of assets in portfolio. Has to be less than number of assets}

\item{ret.fill.na}{Number of NA'sto put in the beginning}

\item{TC}{Tranding costs. If null, then vector without trading cost will be returned}
}
\value{
list
 \item{name mat}{Matrix of returns for top ranked assets}
 \item{name ret}{Vector of returns for tp ranked assets. If TC is not null, tranding cost is in the calculation}
 \item{name rank}{Matrix of ranks for the top}
 \item{name NumberTrades}{Scalar number of trades}
}
\description{
Relative Momentum Strategy
}
