% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sdRollingMatrixCpp}
\alias{sdRollingMatrixCpp}
\title{sdRollingMatrixCpp}
\usage{
sdRollingMatrixCpp(X, period)
}
\arguments{
\item{X}{a matrix of which one want to calculate SD rolling}

\item{period}{number one want to use when taking SD rolling}
}
\description{
SD rolling over specified period with matrix as output
}
