% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SMAMatrixCpp}
\alias{SMAMatrixCpp}
\title{Simpel Moving Average with a Matrix as input}
\usage{
SMAMatrixCpp(X, period)
}
\arguments{
\item{X}{a matrix of which one want to calculate simpel moving average}

\item{period}{number one want to use when taking simpel moving average}
}
\description{
Returns simplle moving average
}
