% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedErr.R
\name{weightedErr}
\alias{weightedErr}
\title{Weighted Error adjusted momentum}
\usage{
weightedErr(x, n = c(1, 3, 6), weights = c(1/3, 1/3, 1/3), period,
  SMAperiod, CompareObj)
}
\arguments{
\item{x}{Object with returns}

\item{n}{Vector in which one choose up to three lookback periods}

\item{weights}{Vector in which one choose up to three weights}

\item{period}{Periods at which one is rebalancing. Can be "days", "weeks", "months", "quarters" or "years"}

\item{SMAperiod}{Period in which one want to adjust prices. Hence this period is used when calculating forecast and thereby errors}

\item{CompareObj}{Object of which final object should be compare to. Ex. if period is monthly, then a frame with the wanted index should be provided}
}
\value{
Weighted Error adjusted momentum of returns
}
\description{
Weighted Error adjusted momentum
}
