% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedROC.R
\name{weightedROC}
\alias{weightedROC}
\title{Weighted ROC}
\usage{
weightedROC(x, n = c(1, 3, 6), weights = c(1/3, 1/3, 1/3))
}
\arguments{
\item{x}{Object with returns}

\item{n}{Vector in which one choose up to three lookback periods}

\item{weights}{Vector in which one choose up to three weights}
}
\value{
Weighted ROC of returns
}
\description{
Weighted ROC
}
