% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weightsPropMatrix}
\alias{weightsPropMatrix}
\title{weightsPropMatrix}
\arguments{
\item{x}{A numeric matrix of t-statistics}

\item{cf}{double confidence level}

\item{k}{int lockback period}
}
\description{
Returns weights given t-statistics and a confidence level
}
