% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeMomentum.R
\name{RelativeMomentum}
\alias{RelativeMomentum}
\title{Relative Momentum Strategy}
\usage{
RelativeMomentum(xts.ret, xts.rank, n = 1, ret.fill.na = 3,
  UseTC = TRUE, TCmin = 30, TCrate = 0.001, size = 1, price)
}
\arguments{
\item{xts.ret}{Object of returns}

\item{xts.rank}{Object of ranked returns}

\item{n}{Number of assets in portfolio. Has to be less than number of assets}

\item{ret.fill.na}{Number of NA'sto put in the beginning}

\item{UseTC}{Logical if one want to use transaction cast}

\item{TCmin}{Minimum transactionamount}

\item{TCrate}{Rate in which the transaction cost is calculated. This is only used if it is over the minimal amount}

\item{size}{Size of long position when one is buying. Increasing this can decrease the size of transaction cost, assuming they are independent of size}

\item{price}{object of prices}
}
\value{
list
 \item{name mat}{Matrix of returns for top ranked assets}
 \item{name ret}{Vector of returns for tp ranked assets. If TC is not null, tranding cost is in the calculation}
 \item{name rank}{Matrix of ranks for the top}
 \item{name NumberTrades}{Scalar number of trades}
}
\description{
Relative Momentum Strategy
}
